requirements.txt
fastapi
uvicorn
whoosh
python-dotenv
google-genai
groq
pydantic